% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip}
\alias{chip}
\title{Chips}
\format{
A data frame with columns:
\describe{
\item{chip_nr}{WCS Chip Number: values from 1 to 330}
\item{grid_row}{WCS Grid Row: values from A to J}
\item{grid_col}{WCS Grid Columns: values from 0 to 40}
\item{grid_coord}{Concatenation of grid_row and grid_col}

}
}
\source{
\url{https://www1.icsi.berkeley.edu/wcs/data/20021219/txt/chip.txt}, downloaded 2021-12-17
}
\usage{
chip
}
\description{
Chip numbers and grid coordinates for the Munsell chart.
}
\examples{
chip

if (require("dplyr")) {

chip \%>\% dplyr::filter(chip_nr > 150)

}
}
\keyword{datasets}
