% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dict}
\alias{dict}
\title{Dictionary}
\format{
A data frame with columns:
\describe{
\item{lang_nr}{WCS Language Number: values from 1 to 110}
\item{term_nr}{Sequential numbering of terms per language}
\item{term}{Transcription of the color term}
\item{term_abb}{A unique abbreviation of the term}
}
}
\source{
\url{https://www1.icsi.berkeley.edu/wcs/data/20041016/txt/dict.txt}, downloaded 2021-12-17
}
\usage{
dict
}
\description{
Dictionary file to lookup term number and the corresponding term and abbreviaton for each language.
}
\keyword{datasets}
