% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{foci_task}
\alias{foci_task}
\title{Foci task}
\usage{
foci_task()
}
\value{
A tibble data frame.
}
\description{
Merge foci data with all the dictionary data (chip, mun_2_lab, dict, lang, speaker)
to obtain one large data frame
}
\examples{
term_task()
}
