% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lang}
\alias{lang}
\title{Languages}
\format{
A data frame with columns:
\describe{
\item{lang_nr}{WCS Language Number: values from 1 to 110}
\item{lang_name}{Name of each language}
\item{lang_country}{Country}
\item{iso_693_3}{ISO code of each language}
}
}
\source{
\url{https://www1.icsi.berkeley.edu/wcs/WCS_SIL_codes.html}, downloaded 2021-12-17
}
\usage{
lang
}
\description{
Information on the WCS language names, including origin and ISO 693-3 codes.
}
\keyword{datasets}
