% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mun_2_lab}
\alias{mun_2_lab}
\title{Munsell to CIEL\emph{a}b}
\format{
A data frame with columns:
\describe{
\item{chip_nr}{WCS Chip Number: values from 1 to 330}
\item{wcs_mv}{WCS code for Munsell Value}
\item{wcs_mh}{WCS code for Munsell Hue}
\item{mun_chroma}{Munsell Chroma}
\item{mun_hue}{Munsell Hue}
\item{mun_value}{Munsell Value}
\item{L_star}{CIE\verb{L*a*b*} \verb{L*} reference value }
\item{a_star}{CIE\verb{L*a*b*} \code{a} reference value}
\item{b_star}{CIE\verb{L*a*b*} \verb{b*} reference value}
}
}
\source{
\url{https://www1.icsi.berkeley.edu/wcs/data/cnum-maps/cnum-vhcm-lab-new.txt}, downloaded 2021-12-17
}
\usage{
mun_2_lab
}
\description{
Mapping table for Munsell Value, Chroma & Hue values to the CIELAB color space.
The conversion from Munsell to CIEL\emph{a}b* was done assuming Illuminant C.
}
\keyword{datasets}
