% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speaker}
\alias{speaker}
\title{Speaker}
\format{
A data frame with columns:
\describe{
\item{lang_nr}{WCS Language Number: values from 1 to 110}
\item{speaker_nr}{Unique number of speakers per language}
\item{speaker_age}{Age, in years}
\item{speaker_sex}{Sex, male or female}
}
}
\source{
\url{https://www1.icsi.berkeley.edu/wcs/data/20100912/spkr-lsas.txt}, downloaded 2021-12-17
}
\usage{
speaker
}
\description{
Information on each speaker of WCS, including age and sex.
}
\keyword{datasets}
