% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{term}
\alias{term}
\title{Term task}
\format{
A data frame with columns:
\describe{
\item{lang_nr}{WCS Language Number: values from 1 to 110}
\item{speaker_nr}{Unique number of speakers per language}
\item{chip_nr}{WCS Chip Number: values from 1 to 330}
\item{term_abb}{A unique abbreviation of the term}
}
}
\source{
\url{https://www1.icsi.berkeley.edu/wcs/data/20021219/txt/term.txt}, downloaded 2021-12-17
}
\usage{
term
}
\description{
Data for the WCS naming task (chip responses).
}
\keyword{datasets}
