% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{term_task}
\alias{term_task}
\title{Term task}
\usage{
term_task()
}
\value{
A tibble data frame.
}
\description{
Merge term data with all the dictionary data (chip, mun_2_lab, dict, lang, speaker)
to obtain one large data frame
}
\examples{
term_task()
}
