% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip}
\alias{chip}
\title{Chip}
\format{
A data frame with columns:
\describe{
\item{chip_nr}{Values from 1 to 330}
\item{grid_row}{Values from A to J}
\item{grid_col}{Values from 0 to 40}
\item{grid_coord}{Concatenation of grid_row and grid_col}

}
}
\source{
\url{https://openflights.org/data.html}, downloaded 2021-12-17
}
\usage{
chip
}
\description{
Useful metadata about airports.
}
\examples{
chip

if (require("dplyr")) {

chip \%>\% dplyr::filter(chip_nr > 150)

}
}
\keyword{datasets}
