% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mun_2_lab}
\alias{mun_2_lab}
\title{Mun2Lab}
\format{
A data frame with columns:
\describe{
\item{faa}{FAA airport code.}
\item{name}{Usual name of the airport.}
\item{lat, lon}{Location of airport.}
\item{alt}{Altitude, in feet.}
\item{tz}{Timezone offset from GMT.}
\item{dst}{Daylight savings time zone. A = Standard US DST: starts on the
second Sunday of March, ends on the first Sunday of November.
U = unknown. N = no dst.}
\item{tzone}{IANA time zone, as determined by GeoNames webservice.}
}
}
\source{
\url{https://openflights.org/data.html}, downloaded 2021-12-17
}
\usage{
mun_2_lab
}
\description{
Useful metadata about airports.
}
\keyword{datasets}
